requirements.txt
# 核心依赖
streamlit>=1.43.2
openai>=1.67.0
duckduckgo_search>=7.5.3
numpy>=1.26.4
requests>=2.32.3
python-dotenv>=1.0.1
FlagEmbedding>=1.3.4
tqdm>=4.67.1
scikit-learn>=1.6.1

# 数据处理相关
pandas>=2.2.3
pyarrow>=19.0.1
regex>=2024.11.6
fpdf2

# 模型相关
torch>=2.6.0
transformers>=4.49.0
sentence-transformers>=3.4.1
tokenizers>=0.21.1

# 基础工具
pydantic>=2.10.6
rich>=13.9.4
python-multipart>=0.0.20